convert database to gamelogic object

+++ app/src/main/java/classes/PlayerState.kt
@@ -1,9 +1,9 @@
 package classes
 
 
-data class PlayerState(var id: Int,
-                        var gameId: Int,
-                       var playerId: Int,
+data class PlayerState(var id: Long,
+                        var gameId: Long,
+                       var playerId: Long,
                        var turn: Int,
                        var drinksTaken: Int,
                        var maxDrinks: Int)
